Attempt to fix GitHub Workflow
On newer 22.04 Ubuntu, it seems the ODBC package has been bumped to "2".

Related commits:
    - fa652e8
    - 6bfd7d2

(cherry picked from commit dbf794dccad5f2865f955e693c85a65bc5046237)

Fix a bunch of ARM32 printf() warnings (Ubuntu 22.04)

(cherry picked from commit b8fdce475d6c8b7a3df3b2ae1b6d7ec64a053d9c)

Attempt #2 to fix GitHub Workflow

Previous commit fixed 22.04 but broke 20.04, so this commit should make
both work.

(cherry picked from commit d7fcd90432494b259b608d5397c6dd2bcdaa3f59)

Fix package typo

(cherry picked from commit 44c995aae67d24f80161096a529ded398bd69c45)

+++ lib/reg/ci.c
@@ -214,7 +214,7 @@ void print_ci(ucontact_info_t *ci)
 	LM_DBG(" ----- UCI DUMP (%p) ------\n", ci);
 	LM_DBG("received: %.*s, path: %.*s\n", ci->received.len, ci->received.s,
 	       ci->path ? ci->path->len : 0, ci->path ? ci->path->s : NULL);
-	LM_DBG("expires: %ld, expires_in: %ld, expires_out: %ld\n", ci->expires,
+	LM_DBG("expires: %lld, expires_in: %ld, expires_out: %ld\n", (long long)ci->expires,
 	       ci->expires_in, ci->expires_out);
 	LM_DBG("q: %d, instance: %.*s, callid: %.*s\n", ci->q, ci->instance.len,
 	       ci->instance.s, ci->callid ? ci->callid->len : 0,
@@ -224,6 +224,6 @@ void print_ci(ucontact_info_t *ci)
 	LM_DBG("user_agent: %.*s, sock: %p, methods: %d\n",
 	       ci->user_agent ? ci->user_agent->len : 0,
 	       ci->user_agent ? ci->user_agent->s : NULL, ci->sock, ci->methods);
-	LM_DBG("last_modified: %ld, attr: %.*s\n", ci->last_modified,
+	LM_DBG("last_modified: %lld, attr: %.*s\n", (long long)ci->last_modified,
 	       ci->attr ? ci->attr->len : 0, ci->attr ? ci->attr->s : NULL);
 }
